+++ src/lib.rs
@@ -17,3 +17,4 @@ pub use self::console::*;
 pub use self::file::FileLoader;
 pub use self::img::*;
 pub use self::input::{InputApi, Keys};
+pub use uni_app::ScanCode;